tests/python/PythonSDK.Tests/requirements.txt
pytest==7.4.2
pytest-mock==3.12.0
langchain==0.0.312
langchain-experimental==0.0.25
openai==0.28.1
pandas==2.1.1
tabulate==0.9.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
pymysql==1.1.0
pyodbc==4.0.39
psycopg2==2.9.9
sqlalchemy==2.0.21
fastapi==0.103.2
tenacity==8.2.3
azure-storage-blob==12.18.2
tiktoken==0.5.1
typing==3.7.4.3
numpy==1.26.0
duckduckgo-search==4.1.0
flask-sqlalchemy==3.1.1
chromadb==0.4.14
unstructured==0.10.22
unstructured[all-docs]==0.10.22
azure-appconfiguration-provider==1.0.0
pylint==3.0.2
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl